refactor(ui): move sort_connections to ui::sorting and details consts to details.rs
Two follow-up cleanups after the tab split:

- sort_connections (75 lines) moves out of main.rs into
  src/ui/sorting.rs. main.rs only orchestrates the loop now;
  ordering connections by column is part of the UI vocabulary
  (it depends on ui::SortColumn).
- DETAIL_LABEL_WIDTH and DETAILS_SPLIT_MIN_WIDTH only have one
  caller (the Details renderer), so they move from ui/mod.rs
  next to the code that uses them.
